Refactored almost everything from ace.py into builder.py; Added compiler version detection
